Refactor: move bare cluster validation to NewCreateClusterLoader
