+++ pkg/eks/api.go
@@ -232,7 +232,12 @@ func ParseConfig(data []byte) (*api.ClusterConfig, error) {
 
 // LoadConfigFromFile loads ClusterConfig from configFile
 func LoadConfigFromFile(configFile string) (*api.ClusterConfig, error) {
-	data, err := readConfig(configFile)
+	return LoadConfigWithReader(configFile, nil)
+}
+
+// LoadConfigWithReader loads ClusterConfig from configFile or configReader.
+func LoadConfigWithReader(configFile string, configReader io.Reader) (*api.ClusterConfig, error) {
+	data, err := readConfig(configFile, configReader)
 	if err != nil {
 		return nil, errors.Wrapf(err, "reading config file %q", configFile)
 	}
@@ -241,12 +246,14 @@ func LoadConfigFromFile(configFile string) (*api.ClusterConfig, error) {
 		return nil, errors.Wrapf(err, "loading config file %q", configFile)
 	}
 	return clusterConfig, nil
-
 }
 
-func readConfig(configFile string) ([]byte, error) {
+func readConfig(configFile string, reader io.Reader) ([]byte, error) {
 	if configFile == "-" {
-		return io.ReadAll(os.Stdin)
+		if reader == nil {
+			reader = os.Stdin
+		}
+		return io.ReadAll(reader)
 	}
 	return os.ReadFile(configFile)
 }
